refactor: split ability classes
