+++ src/entities/abilities/AbstractAbility.py
